+++ lexicmap/cmd/lib-chaining.go
@@ -216,13 +216,14 @@ func (ce *Chainer) Chain(subs *[]*SubstrPair) (*[]*[]int, float64) {
 		first = true
 		for {
 			j = maxscoresIdxs[i] // previous anchor
-			if visited[j] {      // curent anchor is abandoned
-				if len(*path) > 0 {
-					reverseInts(*path)
-					*paths = append(*paths, path)
-					// fmt.Printf("stop at %d, %s\n", i, (*subs)[i])
+			if visited[j] {      // current anchor is abandoned
+				if len(*path) > 0 { // but don't forget already added path
+					// reverseInts(*path)
+					// *paths = append(*paths, path)
+					// // fmt.Printf("stop at %d, %s\n", i, (*subs)[i])
 
-					path = poolChain.Get().(*[]int)
+					// path = poolChain.Get().(*[]int)
+					// *path = (*path)[:0]
 					*path = (*path)[:0]
 				}
 				visited[i] = true // do not check it again