+++ lexicmap/cmd/lib-seq_compare_test.go
@@ -82,7 +82,7 @@ func TestSeqCompare(t *testing.T) {
 		return
 	}
 	if cr != nil {
-		t.Logf("nChains: %d, af: %.2f, identity: %.2f\n", cr.NumChains, cr.AlignedFraction, cr.Identity)
+		t.Logf("nChains: %d, aligned bases: %d, identity: %.2f\n", cr.NumChains, cr.AlignedBases, cr.PIdentity)
 
 		RecycleSeqComparatorResult(cr)
 	}